Apply Factory method
