+++ xchart/src/main/java/org/knowm/xchart/internal/chartpart/Axis.java
@@ -58,6 +58,7 @@ public class Axis<ST extends AxesChartStyler, S extends AxesChartSeries> impleme
   
   private static List<Double> nullList = Collections.emptyList();
   private static int zeroIndex = 0;
+  AxisTickFactory axisTickFactory = new AxisTickCalculatorFactory();
   /**
    * Constructor
    *
@@ -255,7 +256,7 @@ public class Axis<ST extends AxesChartStyler, S extends AxesChartSeries> impleme
       // g.draw(bounds);
 
       // now paint the X-Axis given the above paint zone
-      this.axisTickCalculator = getAxisTickCalculator(bounds.getWidth());
+      getAxisTickCalculator(bounds.getWidth());
       axisTitle.paint(g);
       axisTick.paint(g);
     }
@@ -287,7 +288,7 @@ public class Axis<ST extends AxesChartStyler, S extends AxesChartSeries> impleme
       titleHeight = rectangle.getHeight() + axesChartStyler.getAxisTitlePadding();
     }
 
-    this.axisTickCalculator = getAxisTickCalculator(workingSpace);
+    getAxisTickCalculator(workingSpace);
 
     // Axis tick labels
     double axisTickLabelsHeight = 0.0;
@@ -347,7 +348,7 @@ public class Axis<ST extends AxesChartStyler, S extends AxesChartSeries> impleme
       titleHeight = rectangle.getHeight() + axesChartStyler.getAxisTitlePadding();
     }
 
-    this.axisTickCalculator = getAxisTickCalculator(workingSpace);
+    getAxisTickCalculator(workingSpace);
 
     // Axis tick labels
     double axisTickLabelsHeight = 0.0;
@@ -382,8 +383,7 @@ public class Axis<ST extends AxesChartStyler, S extends AxesChartSeries> impleme
     return titleHeight + axisTickLabelsHeight;
   }
 
-  private AxisTickCalculator_ getAxisTickCalculator(double workingSpace) {
-
+  private void getAxisTickCalculator(double workingSpace) {
     // X-Axis
     if (getDirection() == Direction.X) {
       List<Double> xData = new ArrayList<>();
@@ -420,23 +420,29 @@ public class Axis<ST extends AxesChartStyler, S extends AxesChartSeries> impleme
 
       if (axesChartStyler.getxAxisTickLabelsFormattingFunction() != null) {
     	  if (!xData.isEmpty()) {
-    		  return new AxisTickCalculator_Callback(
+    		  this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Callback,
     	              axesChartStyler.getxAxisTickLabelsFormattingFunction(),
     	              getDirection(),
     	              workingSpace,
     	              min,
     	              max,
     	              nullList,
-    	              axesChartStyler);
+    	              axesChartStyler,
+    	              nullList,
+    	              null,
+    	              zeroIndex);
     	  }
-          return new AxisTickCalculator_Callback(
-              axesChartStyler.getxAxisTickLabelsFormattingFunction(),
-              getDirection(),
-              workingSpace,
-              min,
-              max,
-              xData,
-              axesChartStyler);
+		  this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Callback,
+	              axesChartStyler.getxAxisTickLabelsFormattingFunction(),
+	              getDirection(),
+	              workingSpace,
+	              min,
+	              max,
+	              xData,
+	              axesChartStyler,
+	              nullList,
+	              null,
+	              zeroIndex);
 
       } else if (axesChartStyler instanceof CategoryStyler
           || axesChartStyler instanceof BoxStyler) {
@@ -446,34 +452,85 @@ public class Axis<ST extends AxesChartStyler, S extends AxesChartSeries> impleme
             (AxesChartSeriesCategory) chart.getSeriesMap().values().iterator().next();
         List<?> categories = (List<?>) axesChartSeries.getXData();
         DataType axisType = chart.getAxisPair().getXAxis().getDataType();
-
-        return new AxisTickCalculator_Category(
-            getDirection(), workingSpace, categories, axisType, axesChartStyler);
+        this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Category,
+	              null,
+	              getDirection(),
+	              workingSpace,
+	              Double.NaN,
+	              Double.NaN,
+	              nullList,
+	              axesChartStyler,
+	              categories,
+	              axisType,
+	              zeroIndex);
 
       } else if (getDataType() == Series.DataType.Date
           && !(axesChartStyler instanceof HeatMapStyler)) {
-
-        return new AxisTickCalculator_Date(getDirection(), workingSpace, min, max, axesChartStyler);
+    	  this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Date,
+	              null,
+	              getDirection(),
+	              workingSpace,
+	              min,
+	              max,
+	              nullList,
+	              axesChartStyler,
+	              nullList,
+	              null,
+	              zeroIndex);
 
       } else if (axesChartStyler.isXAxisLogarithmic()) {
-
-        return new AxisTickCalculator_Logarithmic(
-            getDirection(), workingSpace, min, max, axesChartStyler);
+    	  this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Logarithmic,
+	              null,
+	              getDirection(),
+	              workingSpace,
+	              min,
+	              max,
+	              nullList,
+	              axesChartStyler,
+	              nullList,
+	              null,
+	              zeroIndex);
 
       } else if (axesChartStyler instanceof HeatMapStyler) {
 
         List<?> categories = (List<?>) ((HeatMapChart) chart).getHeatMapSeries().getXData();
         DataType axisType = chart.getAxisPair().getXAxis().getDataType();
-
-        return new AxisTickCalculator_Category(
-            getDirection(), workingSpace, categories, axisType, axesChartStyler);
+        this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Category,
+	              null,
+	              getDirection(),
+	              workingSpace,
+	              Double.NaN,
+	              Double.NaN,
+	              nullList,
+	              axesChartStyler,
+	              categories,
+	              axisType,
+	              zeroIndex);
       } else {
         if (!xData.isEmpty()) {
-          return new AxisTickCalculator_Number(
-              getDirection(), workingSpace, min, max, xData, axesChartStyler, zeroIndex);
+        	this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Number,
+  	              null,
+  	              getDirection(),
+  	              workingSpace,
+  	              min,
+  	              max,
+  	              xData,
+  	              axesChartStyler,
+  	              nullList,
+  	              null,
+  	              zeroIndex);
         }
-        return new AxisTickCalculator_Number(
-            getDirection(), workingSpace, min, max, nullList, axesChartStyler, zeroIndex);
+        this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Number,
+	              null,
+	              getDirection(),
+	              workingSpace,
+	              min,
+	              max,
+	              nullList,
+	              axesChartStyler,
+	              nullList,
+	              null,
+	              zeroIndex);
       }
     }
 
@@ -512,42 +569,83 @@ public class Axis<ST extends AxesChartStyler, S extends AxesChartSeries> impleme
 
       if (axesChartStyler.getyAxisTickLabelsFormattingFunction() != null) {
     	  if (!yData.isEmpty()) {
-    		  return new AxisTickCalculator_Callback(
+    		  this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Callback,
     	              axesChartStyler.getxAxisTickLabelsFormattingFunction(),
     	              getDirection(),
     	              workingSpace,
     	              min,
     	              max,
     	              nullList,
-    	              axesChartStyler);
+    	              axesChartStyler,
+    	              nullList,
+    	              null,
+    	              zeroIndex);
     	  }
-          return new AxisTickCalculator_Callback(
-              axesChartStyler.getxAxisTickLabelsFormattingFunction(),
-              getDirection(),
-              workingSpace,
-              min,
-              max,
-              yData,
-              axesChartStyler);
+    	  this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Callback,
+	              axesChartStyler.getxAxisTickLabelsFormattingFunction(),
+	              getDirection(),
+	              workingSpace,
+	              min,
+	              max,
+	              yData,
+	              axesChartStyler,
+	              nullList,
+	              null,
+	              zeroIndex);
 
       } else if (axesChartStyler.isYAxisLogarithmic() && getDataType() != Series.DataType.Date) {
 
-        return new AxisTickCalculator_Logarithmic(
-            getDirection(), workingSpace, min, max, axesChartStyler, getYIndex());
+    	  this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Logarithmic,
+	              null,
+	              getDirection(),
+	              workingSpace,
+	              min,
+	              max,
+	              nullList,
+	              axesChartStyler,
+	              nullList,
+	              null,
+	              zeroIndex);
       } else if (axesChartStyler instanceof HeatMapStyler) {
 
         List<?> categories = (List<?>) ((HeatMapChart) chart).getHeatMapSeries().getYData();
         DataType axisType = chart.getAxisPair().getYAxis().getDataType();
-
-        return new AxisTickCalculator_Category(
-            getDirection(), workingSpace, categories, axisType, axesChartStyler);
+        this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Category,
+	              null,
+	              getDirection(),
+	              workingSpace,
+	              Double.NaN,
+	              Double.NaN,
+	              nullList,
+	              axesChartStyler,
+	              categories,
+	              axisType,
+	              zeroIndex);
       } else {
         if (!yData.isEmpty()) {
-          return new AxisTickCalculator_Number(
-              getDirection(), workingSpace, min, max, yData, axesChartStyler, zeroIndex);
+        	this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Number,
+  	              null,
+  	              getDirection(),
+  	              workingSpace,
+  	              min,
+  	              max,
+  	              yData,
+  	              axesChartStyler,
+  	              nullList,
+  	              null,
+  	              zeroIndex);
         }
-        return new AxisTickCalculator_Number(
-            getDirection(), workingSpace, min, max, nullList, axesChartStyler, getYIndex());
+        this.axisTickCalculator = axisTickFactory.calculator(AxisTickCalculatorType.AxisTickCalculator_Number,
+	              null,
+	              getDirection(),
+	              workingSpace,
+	              min,
+	              max,
+	              nullList,
+	              axesChartStyler,
+	              nullList,
+	              null,
+	              getYIndex());
       }
     }
   }
