+++ xchart/src/main/java/org/knowm/xchart/internal/chartpart/AxisTickCalculator_Category.java
@@ -9,6 +9,7 @@ import java.util.List;
 import org.knowm.xchart.internal.Utils;
 import org.knowm.xchart.internal.chartpart.Axis.Direction;
 import org.knowm.xchart.internal.series.Series;
+import org.knowm.xchart.internal.series.Series.DataType;
 import org.knowm.xchart.style.AxesChartStyler;
 
 /**
@@ -19,6 +20,9 @@ import org.knowm.xchart.style.AxesChartStyler;
  */
 class AxisTickCalculator_Category extends AxisTickCalculator_ {
   private static List<Double> nullList = Collections.emptyList();
+  private List<?> categories;
+  private Series.DataType getAxisType;
+  private DataType axisType;
   /**
    * Constructor
    *
@@ -36,12 +40,13 @@ class AxisTickCalculator_Category extends AxisTickCalculator_ {
       AxesChartStyler styler) {
 
     super(axisDirection, workingSpace, Double.NaN, Double.NaN, nullList, styler);
-
-    calculate(categories, axisType);
+    this.categories = categories;
+    this.getAxisType = axisType;
+    calculate();
   }
 
-  private void calculate(List<?> categories, Series.DataType axisType) {
-
+  @Override
+  protected void calculate() {
     // tick space - a percentage of the working space available for ticks
     double tickSpace = styler.getPlotContentSize() * workingSpace; // in plot space
     // System.out.println("workingSpace: " + workingSpace);
